training debugging

+++ neural_certificate_examples/controllers/neural_clf_controller.py
@@ -187,7 +187,7 @@ class NeuralCLFController(pl.LightningModule, CLFController):
         P = self.dynamics_model.P.type_as(x)
         P = torch.eye(self.dynamics_model.n_dims)
         P = P.reshape(1, self.dynamics_model.n_dims, self.dynamics_model.n_dims)
-        x_goal = self.dynamics_model.goal_point
+        x_goal = self.dynamics_model.goal_point.type_as(x)
         V = V + 0.5 * F.bilinear(x - x_goal, x - x_goal, P).squeeze()
         P = P.reshape(self.dynamics_model.n_dims, self.dynamics_model.n_dims)
         JV = JV + F.linear(x - x_goal, P).reshape(
